requirements.txt
flask==2.0.1
numpy
matplotlib
Pillow==8.2.0
Flask-WTF==0.15.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
werkzeug==2.0.3
sqlalchemy==1.4.46
gunicorn==20.1.0 